requirements.txt
beautifulsoup4~=4.10.0
requests~=2.26.0
github3.py~=2.0.0